# Use Cow<str> for faster serde decoding, update some deps