--- a/rust/Cargo.toml
+++ b/rust/Cargo.toml
@@ -17,7 +17,7 @@ num-derive = "0.2"
 num-traits = "0.2"
 protobuf = "2.18"
 rmp-serde = "0.15.0"
+rmpv = { "version" = "0.4.6", "features" = ["with-serde"] }
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
-rmpv = "0.4.6"