This changes Log's `String` fields to `Cow<str>` (in rust), so the values can be borrowed from the buffer when deserializing, rather than copied. The result is a pretty significant speed boost for the serde deserializers, at the cost of some verbosity in the Log structs.

before:
```
test rmp_serde_deserialize             ... bench:       2,353 ns/iter (+/- 532) = 121 MB/s
test rust_bincode_deserialize          ... bench:       1,830 ns/iter (+/- 249) = 218 MB/s
test serde_json_deserialize            ... bench:       2,858 ns/iter (+/- 422) = 211 MB/s
```
after:
```
test rmp_serde_deserialize             ... bench:         531 ns/iter (+/- 141) = 540 MB/s
test rust_bincode_deserialize          ... bench:         320 ns/iter (+/- 73) = 1250 MB/s
test serde_json_deserialize            ... bench:       1,212 ns/iter (+/- 191) = 499 MB/s
```

I also added some `_dynamic` benchmarks for rmp and serde_json which de/serialize from/to the libraries' respective dynamic `Value` enums, to satisfy my curiosity.
```
test rmp_serde_deserialize_dynamic     ... bench:       1,724 ns/iter (+/- 1,031) = 166 MB/s
test rmp_serde_serialize_dynamic       ... bench:         421 ns/iter (+/- 83) = 681 MB/s
test serde_json_deserialize_dynamic    ... bench:       9,390 ns/iter (+/- 1,925) = 64 MB/s
test serde_json_serialize_dynamic      ... bench:       1,459 ns/iter (+/- 214) = 414 MB/s
```